refactor(service): improve strategy pattern for PredictionService class (#44) (#45)
* refactor(service): improve strategy pattern for PredictionService class
* create a new strategy for model selection
* extract PredictionService class from CADOCS
* modify LanguageModel class to an abstract class
* refactor(service): remove unused import
* remove unused CADOCS import

Closes #42
